Move CommandError under datalad.runner.exception
This is the core exception type used, and it is only raised by the runner
(except for test code).

Reimport to its previous place to avoid breakage.
