+++ datalad/support/exceptions.py
@@ -11,12 +11,12 @@
 
 import logging
 import re
-import sys
 import traceback
 from os import linesep
 from pathlib import Path
 from pprint import pformat
 
+from datalad.runner.exception import CommandError
 
 lgr = logging.getLogger('datalad.support.exceptions')
 
@@ -179,68 +179,6 @@ def _format_json_error_messages(recs):
             for m, n in msgs.items()
         )
     )
-
-
-class CommandError(RuntimeError):
-    """Thrown if a command call fails.
-
-    Note: Subclasses should override `to_str` rather than `__str__` because
-    `to_str` is called directly in datalad.cmdline.main.
-    """
-
-    def __init__(self, cmd="", msg="", code=None, stdout="", stderr="", cwd=None,
-                 **kwargs):
-        RuntimeError.__init__(self, msg)
-        self.cmd = cmd
-        self.msg = msg
-        self.code = code
-        self.stdout = stdout
-        self.stderr = stderr
-        self.cwd = cwd
-        self.kwargs = kwargs
-
-    def to_str(self, include_output=True):
-        from datalad.utils import (
-            ensure_unicode,
-            join_cmdline,
-        )
-        to_str = "{}: ".format(self.__class__.__name__)
-        cmd = self.cmd
-        if cmd:
-            to_str += "'{}'".format(
-                # go for a compact, normal looking, properly quoted
-                # command rendering if the command is in list form
-                join_cmdline(cmd) if isinstance(cmd, list) else cmd
-            )
-        if self.code:
-            to_str += " failed with exitcode {}".format(self.code)
-        if self.cwd:
-            # only if not under standard PWD
-            to_str += " under {}".format(self.cwd)
-        if self.msg:
-            # typically a command error has no specific idea
-            to_str += " [{}]".format(ensure_unicode(self.msg))
-
-        if self.kwargs:
-            to_str += " [info keys: {}]".format(
-                ', '.join(self.kwargs.keys()))
-
-            if 'stdout_json' in self.kwargs:
-                to_str += _format_json_error_messages(
-                    self.kwargs['stdout_json'])
-
-        if not include_output:
-            return to_str
-
-        if self.stdout:
-            to_str += " [out: '{}']".format(ensure_unicode(self.stdout).strip())
-        if self.stderr:
-            to_str += " [err: '{}']".format(ensure_unicode(self.stderr).strip())
-
-        return to_str
-
-    def __str__(self):
-        return self.to_str()
 
 
 class MissingExternalDependency(RuntimeError):
